requirements.txt
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.11
cycler==0.11.0
Deprecated==1.2.13
fonttools==4.32.0
google-auth==2.6.0
idna==3.3
joblib==1.1.0
kiwisolver==1.4.2
kubernetes==21.7.0
llvmlite==0.38.0
matplotlib==3.5.1
numba==0.55.1
numpy==1.21.5
oauthlib==3.2.0
packaging==21.3
pandas==1.4.2
patsy==0.5.2
Pillow==9.1.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pymannkendall==1.4.2
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
ruptures==1.1.6
scikit-learn==1.0.2
scipy==1.8.0
six==1.16.0
sktime==0.11.3
statsmodels==0.13.2
threadpoolctl==3.1.0
urllib3==1.26.9
websocket-client==1.2.3
wrapt==1.14.0
